splitting into multiple files for drone package
