+++ internal/agents/drone/movement.go
