+++ internal/agents/drone/shared-data.go
